v5第三次提交

+++ yudao-module-steam/yudao-module-steam-biz/src/main/java/cn/iocoder/yudao/module/steam/service/fin/v5/res/V5ProductBuyRes.java
@@ -5,6 +5,8 @@ import lombok.Data;
 import lombok.Getter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+
 @Data
 public class V5ProductBuyRes implements Serializable {
     private int code;
@@ -12,7 +14,7 @@ public class V5ProductBuyRes implements Serializable {
     private PaymentData data;
     @Data
     public static class PaymentData {
-        private int payAmount;
+        private BigDecimal payAmount;
         private ReceiverInfo receiverInfo;
         private String merchantOrderNo;
         private String orderNo;
